+++ include/game/components/bullethell/bullet.h
@@ -4,7 +4,7 @@ namespace Game::BulletHell
 {
     struct Bullet
     {
-        float telegraph_time;
+        int telegraph_time;
         bool is_active;
         bool is_damageable;
         int damage;
